reorganises the Select-tags custom styles component code
