+++ src/app/components/select-tags/Select-tags.jsx
@@ -1,6 +1,7 @@
 import React, { Component } from "react";
 import Select from "react-select";
 import PropTypes from "prop-types";
+import { customStyles } from "./selectTagsCustomStyle";
 
 const propTypes = {
     id: PropTypes.string,
@@ -19,45 +20,6 @@ const propTypes = {
     handleChange: PropTypes.func,
     multipleSelection: PropTypes.bool,
     error: PropTypes.string,
-};
-
-
-const customStyles = {
-    option: (styles, state) => ({
-        ...styles,
-        color: "black",
-        backgroundColor: state.isDisabled ? undefined : state.isSelected ? "#ddd" : state.isFocused ? "#ddd" : "white",
-        ":active": {
-            ...styles[":active"],
-            backgroundColor: !state.isDisabled ? (state.isSelected ? "#ddd" : "#ddd") : undefined,
-        },
-        padding: 20,
-    }),
-    container: (styles, state) => ({
-        ...styles,
-        border: "3px solid #58585B",
-        borderRadius: "4px",
-        boxShadow: state.isFocused ? "0 0 0 3px #ffbf47" : "none",
-        backgroundColor: "#FFFFFF",
-        outlineStyle: "none",
-    }),
-    control: styles => ({
-        ...styles,
-        border: "none",
-        backgroundColor: "transparent",
-        outlineStyle: "none",
-    }),
-    menuList: styles => ({
-        ...styles,
-        marginBottom: "35px",
-    }),
-    multiValueRemove: styles => ({
-        ...styles,
-        ":hover": {
-            backgroundColor: "#ddd",
-            color: "#333333",
-        },
-    }),
 };
 
 class SelectTags extends Component {
